code-examples/dj-scratch/requirements.txt
#
# This file is autogenerated by pip-compile with python 3.10
# To update, run:
#
#    pip-compile
#
asgiref==3.4.1
    # via django
attrs==21.2.0
    # via pytest
black==21.11b1
    # via -r requirements.in
click==8.0.3
    # via
    #   black
    #   pip-tools
django==3.2.9
    # via
    #   -r requirements.in
    #   django-extensions
    #   django-stubs
    #   django-stubs-ext
    #   django-zen-queries
    #   djangorestframework
django-extensions==3.1.5
    # via -r requirements.in
django-stubs==1.9.0
    # via -r requirements.in
django-stubs-ext==0.3.1
    # via
    #   -r requirements.in
    #   django-stubs
django-zen-queries==2.1.0
    # via -r requirements.in
djangorestframework==3.12.4
    # via -r requirements.in
iniconfig==1.1.1
    # via pytest
isort==5.10.1
    # via -r requirements.in
mypy==0.910
    # via
    #   -r requirements.in
    #   django-stubs
mypy-extensions==0.4.3
    # via
    #   black
    #   mypy
packaging==21.3
    # via pytest
pathspec==0.9.0
    # via black
pep517==0.12.0
    # via pip-tools
pip-tools==6.4.0
    # via -r requirements.in
platformdirs==2.4.0
    # via black
pluggy==1.0.0
    # via pytest
psycopg2==2.9.2
    # via -r requirements.in
py==1.11.0
    # via pytest
pyparsing==3.0.6
    # via packaging
pytest==6.2.5
    # via
    #   -r requirements.in
    #   pytest-django
pytest-django==4.4.0
    # via -r requirements.in
pytz==2021.3
    # via django
regex==2021.11.10
    # via black
sqlparse==0.4.2
    # via django
toml==0.10.2
    # via
    #   django-stubs
    #   mypy
    #   pytest
tomli==1.2.2
    # via
    #   black
    #   pep517
types-pytz==2021.3.1
    # via django-stubs
types-pyyaml==6.0.1
    # via django-stubs
typing-extensions==4.0.0
    # via
    #   black
    #   django-stubs
    #   django-stubs-ext
    #   mypy
wheel==0.37.0
    # via pip-tools

# The following packages are considered to be unsafe in a requirements file:
# pip
# setuptools
